may less memory

+++ src/kvsrv/common.go
@@ -8,7 +8,7 @@ type PutAppendArgs struct {
 	// Field names must start with capital letters,
 	// otherwise RPC will break.
 	ClientId   int64
-	SequenceId int64
+	SequenceId int
 }
 
 type PutAppendReply struct {
@@ -19,7 +19,7 @@ type GetArgs struct {
 	Key string
 	// You'll have to add definitions here.
 	ClientId   int64
-	SequenceId int64
+	SequenceId int
 }
 
 type GetReply struct {
